requirements.txt
Django==1.11.6
django-model-utils==3.0.0
django-notifications-hq==1.3
jsonfield==2.0.2
pkg-resources==0.0.0
pytz==2017.2
